extract login function into a single go file
